moves common attributes of input tag field in InputField class
